kyc_backend/requirements.txt
Flask==3.0.0
pypyodbc==1.3.6
cryptography==41.0.7
argon2-cffi==23.1.0
PyJWT==2.8.0
reportlab==4.0.8
python-dotenv==1.0.0
flask-cors==4.0.0
